feat: encapsulamento de classes
