+++ tabs/__init__.py
@@ -1,0 +1,7 @@
+# Este arquivo permite que a pasta tabs funcione como um pacote Python
+from .geral import render_geral
+from .aspectos_sociais import render_aspectos_sociais
+from .desempenho import render_desempenho
+
+# Isso permite importar diretamente:
+# from tabs import render_geral, render_aspectos_sociais, render_desempenho
