+++ utils/__init__.py
@@ -1,0 +1,6 @@
+# Este arquivo permite que a pasta utils funcione como um pacote Python
+from .data_loader import load_data
+from .mappings import get_mappings
+
+# Isso permite importar diretamente:
+# from utils import load_data, get_mappings
